promote candidate impl to be called in prod code
