--- .github/workflows/set-benchmark-experiments.yml
@@ -1,111 +1,0 @@
-name: Set Existing-Index A/B (opt-in)
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ahnlich/**
-      - benchmarks/**
-      - .github/workflows/set-benchmark-experiments.yml
-    types: [opened, reopened, synchronize, labeled]
-  workflow_dispatch:
-    inputs:
-      control_ref:
-        description: Control revision (build without candidate feature)
-        default: main
-        required: true
-      batches:
-        description: Space-separated Set batch sizes
-        default: "100 1000"
-        required: true
-      concurrency:
-        description: Space-separated concurrency levels
-        default: "1 4"
-        required: true
-      repeats:
-        description: Repeats with alternating A/B order
-        default: "3"
-        required: true
-      seconds:
-        description: Seconds per measured update run
-        default: "10"
-        required: true
-
-permissions:
-  contents: read
-
-env:
-  CARGO_TERM_COLOR: always
-  GHZ_VERSION: v0.121.0
-
-jobs:
-  set-ab:
-    if: github.event_name == 'workflow_dispatch' || contains(github.event.pull_request.labels.*.name, 'set-benchmark-experiments')
-    runs-on: ubuntu-latest
-    timeout-minutes: 120
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          path: candidate
-          ref: ${{ github.event.pull_request.head.sha || github.sha }}
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ inputs.control_ref || github.event.pull_request.base.sha }}
-          path: control
-      - uses: Swatinem/rust-cache@v2
-        with:
-          workspaces: |
-            candidate/ahnlich
-            candidate/benchmarks
-            control/ahnlich
-      - name: Install protoc and ghz
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y protobuf-compiler
-          curl -fsSL -o ghz.tar.gz \
-            "https://github.com/bojand/ghz/releases/download/${GHZ_VERSION}/ghz-linux-x86_64.tar.gz"
-          tar -xzf ghz.tar.gz ghz
-          sudo install -m 755 ghz /usr/local/bin/ghz
-      - name: Build control
-        working-directory: control/ahnlich
-        run: cargo build --release --locked --bin ahnlich-db
-      - name: Build candidate
-        working-directory: candidate/ahnlich
-        run: cargo build --release --locked --bin ahnlich-db --features bench-existing-predicate-index
-      - name: Build fixture helper
-        working-directory: candidate/benchmarks
-        run: cargo build --release --locked --bin setup_set
-      - name: Run Set A/B
-        env:
-          BATCHES: ${{ inputs.batches || '100 1000' }}
-          CONCURRENCY: ${{ inputs.concurrency || '1 4' }}
-          REPEATS: ${{ inputs.repeats || '3' }}
-          SECONDS_PER_RUN: ${{ inputs.seconds || '10' }}
-        shell: bash
-        run: |
-          read -r -a batch_args <<< "$BATCHES"
-          read -r -a concurrency_args <<< "$CONCURRENCY"
-          python3 candidate/benchmarks/run_set_ab.py \
-            --control-bin "$GITHUB_WORKSPACE/control/ahnlich/target/release/ahnlich-db" \
-            --candidate-bin "$GITHUB_WORKSPACE/candidate/ahnlich/target/release/ahnlich-db" \
-            --setup-bin "$GITHUB_WORKSPACE/candidate/benchmarks/target/release/setup_set" \
-            --control-ref "$(git -C control rev-parse HEAD)" \
-            --candidate-ref "$(git -C candidate rev-parse HEAD)" \
-            --batches "${batch_args[@]}" --concurrency "${concurrency_args[@]}" \
-            --repeats "$REPEATS" --seconds "$SECONDS_PER_RUN" \
-            --cases 0:100 1:100 4:100 \
-            --results "$GITHUB_WORKSPACE/set-results"
-      - name: Publish summary
-        if: always()
-        run: |
-          if [ -f set-results/SUMMARY.md ]; then
-            cat set-results/SUMMARY.md >> "$GITHUB_STEP_SUMMARY"
-          else
-            echo 'Sweep did not complete; inspect the uploaded logs.' >> "$GITHUB_STEP_SUMMARY"
-          fi
-      - uses: actions/upload-artifact@v4
-        if: always()
-        with:
-          name: set-existing-index-ab-${{ github.run_number }}
-          path: set-results
-          retention-days: 30
