+++ ahnlich/db/src/engine/store.rs
@@ -1248,15 +1248,11 @@ impl Store {
             .collect();
 
         if let Some(predicate_insert) = predicate_insert {
-            #[cfg(feature = "bench-existing-predicate-index")]
             self.predicate_indices.add_existing_index_candidate(
                 predicate_insert,
                 parallelism_config,
                 active_requests,
             );
-            #[cfg(not(feature = "bench-existing-predicate-index"))]
-            self.predicate_indices
-                .add(predicate_insert, parallelism_config, active_requests);
         }
 
         if !self.non_linear_indices.is_empty() {
