+++ benchmarks/SEARCH_BENCHMARKS.md
@@ -2,8 +2,8 @@
 
 Measures end-to-end QPS for ahnlich-db under concurrent gRPC load.
 
-For the existing-predicate-index **Set control/candidate experiment**, see
-[SET_BENCHMARKS.md](SET_BENCHMARKS.md). It uses a separate ghz runner and prebuilt binaries.
+For the `Set` benchmark, see [SET_BENCHMARKS.md](SET_BENCHMARKS.md). It uses deterministic
+synthetic fixtures and a separate ghz runner.
 
 ## Requirements
 
@@ -17,13 +17,19 @@ installation belongs in `../ahnlich/similarity/sift_1m/`. It can be downloaded f
 ## Running
 
 ```bash
-./run_baseline.sh
+./run_getsimn.sh
 ```
 
 Run against the local SIFT1M installation with:
 
 ```bash
-./run_baseline.sh --sift-1m
+./run_getsimn.sh --sift-1m
+```
+
+Run a prebuilt server binary with:
+
+```bash
+SERVER_BIN=/path/to/ahnlich-db ./run_getsimn.sh
 ```
 
 `SIFT_DIR` remains available when the dataset lives elsewhere.
@@ -32,7 +38,7 @@ Builds the release binaries, starts a server, loads the data, runs the sweep, wr
 summary. Paths resolve relative to the script. The server is stopped on exit, including
 on failure and Ctrl-C.
 
-Output lands in `results/baseline_<timestamp>/`:
+Output lands in `results/getsimn_<timestamp>/`:
 
 ```
 SUMMARY.md                  results table
@@ -123,7 +129,7 @@ rows.
 | `RESULTS_DIR` | timestamped | |
 
 ```bash
-TOTAL_REQUESTS=50000 CONCURRENCY_LEVELS="1 4 8 16 32" ./run_baseline.sh
+TOTAL_REQUESTS=50000 CONCURRENCY_LEVELS="1 4 8 16 32" ./run_getsimn.sh
 ```
 
 ## Reading the results
